Update Invoke-ADGenerator.ps1
Fixed part of the Script where dc=mayorsec is hardcoded instead of pulling from the $domainFront variable that was created & populated.  I provide some spacing for readability as well as putting the $domainFront and $domainBack at the top of the AddADUser Function.  NOTE: I replaced the dc=local with dc=$domainBack aswell